Generator/Client/requirements.txt
google-cloud-pubsub
google-auth